fadiga2

+++ ApiQuiz/src/main/java/com/quizapi/ApiQuiz/modele/Domaine.java
@@ -15,6 +15,6 @@ import java.util.List;
 public class Domaine {
     @Id
     @GeneratedValue (strategy = GenerationType.IDENTITY)
-    private int idDomaine;
+    private Long idDomaine;
     private String nom;
 }
